requirement.txt
psycopg2-binary==2.9.9
python-dotenv==1.1.1
pandas==2.1.1
tabulate==0.9.0
